installer: vendor the driver, automate its inputs, and fix reinstall/restore
Driver vendoring
- driver/ is now a verbatim copy of upstream omen-fan-control's hp-wmi
  tree (unchanged, still C/GPL-2), so the installer works without a
  separate checkout. Nothing in Pyren ever writes to it - stage-source
  now runs before patch-source, so patching touches the /usr/src copy
  and the vendored tree stays a pristine snapshot.

Automatic install mode
- installer.autodetect reads what an install needs instead of asking:
  board id/model from DMI, whether the board is already in the driver's
  own hp_wmi_feature_boards table, fan ceilings from the last
  calibration. installer.apply accepts auto: true to fill in whatever a
  request leaves unset.
- The board-params variant is decided, not guessed: all four variants
  share one fan profile and a board already on the OMEN/Victus
  thermal-profile path never has its variant read at all, so most boards
  get params_effect: inert and no caveat. Where it is live, EC offsets
  0x59/0x95 are read via ec_sys (read-only) and whichever holds a value
  the OMEN v1 profile recognises names the variant.
- The app offers automatic and manual as two explicit modes rather than
  a button over a form; manual fields are not sent while off screen.
- Optional plan steps (initramfs regen, unloading a module that may not
  be loaded, etc.) get a checkbox each; required steps have none, since
  the daemon refuses to skip them.

Bug fixes found by actually running the installer on 8D2F
- restore-backups only renamed the .bak back and never removed the
  module the install had added, so distros that ship it compressed
  (hp-wmi.ko.zst) ended up with both on disk and depmod kept resolving
  to ours - "restore" reported success while doing nothing. It now
  removes the leftover, and also cleans up the two-modules-with-no-backup
  state that earlier runs left behind.
- installer.plan's "fan control already works, nothing to install"
  warning didn't distinguish a stock driver that supports the board from
  Pyren's own patched driver already doing the work; the latter now gets
  a distinct warning and is not blocked, so a reinstall (e.g. after a
  kernel upgrade with no hook) is not refused as "already working".
- The kernel-log diagnostic warned on every hp-wmi dmesg line, including
  a driver announcing it loaded fine, so a healthy machine got a warning
  with nothing wrong in it. It now classifies lines and explains the
  expected "module verification failed / tainting kernel" notice that
  follows installing any out-of-tree module.
- The "kernel already has fan control" warning fired even when this
  board wasn't recognised by the stock driver (no pwm1) - the one case
  the patch is actually for - and called installing "probably
  unnecessary" right under a verdict saying the opposite. Split into two
  warnings keyed on whether hp-wmi is loaded.
- Six message literals across the daemon had lost their line-continuation
  backslash and were leaking source indentation into user-facing text.
  Added a workspace-wide test (pyren_core::msg) that scans every string
  literal for both a swallowed continuation and one glued to the next
  word, distinguishing them from deliberate alignment by position rather
  than pattern-matching the fix.

Docs (FINDINGS.md, TODO.md, README, docs/01-ipc-protocol.md,
docs/03-frontend.md) updated to match: the driver has been installed and
verified working on the test laptop, board-params is measured rather
than guessed, and the once-never-run installer execution path is no
longer that.

+++ daemon/crates/hotkey/src/devices.rs
@@ -633,11 +633,11 @@ mod tests {
         let presses = device.absorb(&bytes(&[
             // Ctrl down: scancode, keycode, sync. Reports nothing.
             (EV_MSC, MSC_SCAN, 0x1d),
-            (EV_KEY, KEY_LEFTCTRL as u16, 1),
+            (EV_KEY, KEY_LEFTCTRL, 1),
             (EV_SYN, SYN_REPORT, 0),
             // Alt down.
             (EV_MSC, MSC_SCAN, 0x38),
-            (EV_KEY, KEY_LEFTALT as u16, 1),
+            (EV_KEY, KEY_LEFTALT, 1),
             (EV_SYN, SYN_REPORT, 0),
             // P down: the only press in the whole sequence.
             (EV_MSC, MSC_SCAN, 0x19),
@@ -657,8 +657,8 @@ mod tests {
     #[test]
     fn releasing_a_modifier_stops_it_being_held() {
         let mut device = device();
-        device.absorb(&bytes(&[(EV_KEY, KEY_LEFTCTRL as u16, 1), (EV_SYN, SYN_REPORT, 0)]));
-        device.absorb(&bytes(&[(EV_KEY, KEY_LEFTCTRL as u16, 0), (EV_SYN, SYN_REPORT, 0)]));
+        device.absorb(&bytes(&[(EV_KEY, KEY_LEFTCTRL, 1), (EV_SYN, SYN_REPORT, 0)]));
+        device.absorb(&bytes(&[(EV_KEY, KEY_LEFTCTRL, 0), (EV_SYN, SYN_REPORT, 0)]));
         let presses = device.absorb(&bytes(&[(EV_KEY, 25, 1), (EV_SYN, SYN_REPORT, 0)]));
 
         assert_eq!(presses.len(), 1);
@@ -670,8 +670,8 @@ mod tests {
     #[test]
     fn a_held_modifier_repeating_is_still_held() {
         let mut device = device();
-        device.absorb(&bytes(&[(EV_KEY, KEY_LEFTSHIFT as u16, 1), (EV_SYN, SYN_REPORT, 0)]));
-        device.absorb(&bytes(&[(EV_KEY, KEY_LEFTSHIFT as u16, 2), (EV_SYN, SYN_REPORT, 0)]));
+        device.absorb(&bytes(&[(EV_KEY, KEY_LEFTSHIFT, 1), (EV_SYN, SYN_REPORT, 0)]));
+        device.absorb(&bytes(&[(EV_KEY, KEY_LEFTSHIFT, 2), (EV_SYN, SYN_REPORT, 0)]));
         let presses = device.absorb(&bytes(&[(EV_KEY, 25, 1), (EV_SYN, SYN_REPORT, 0)]));
 
         assert!(presses[0].modifiers.shift);
@@ -682,11 +682,11 @@ mod tests {
     #[test]
     fn the_two_sides_of_a_modifier_are_the_same_modifier() {
         let mut left = device();
-        left.absorb(&bytes(&[(EV_KEY, KEY_LEFTSHIFT as u16, 1), (EV_SYN, SYN_REPORT, 0)]));
+        left.absorb(&bytes(&[(EV_KEY, KEY_LEFTSHIFT, 1), (EV_SYN, SYN_REPORT, 0)]));
         let from_left = left.absorb(&bytes(&[(EV_KEY, 25, 1), (EV_SYN, SYN_REPORT, 0)]));
 
         let mut right = device();
-        right.absorb(&bytes(&[(EV_KEY, KEY_RIGHTSHIFT as u16, 1), (EV_SYN, SYN_REPORT, 0)]));
+        right.absorb(&bytes(&[(EV_KEY, KEY_RIGHTSHIFT, 1), (EV_SYN, SYN_REPORT, 0)]));
         let from_right = right.absorb(&bytes(&[(EV_KEY, 25, 1), (EV_SYN, SYN_REPORT, 0)]));
 
         assert_eq!(from_left[0].modifiers, from_right[0].modifiers);
